Create guess_no_game.java
This is the 1st Task of CodSoft Internship! The 1st task is about Number Guessing Game using

Java Programming Language. A program generates a random number and asks the user to guess it and Provides feedback on whether the guess is too high or too low until the user guesses the correct number. Maximum attempts are 10 if the user Guesses the correct number, the Score will be incremented by 1, then Program ask the user to play the game again or display the total score of the user.

#codsoft #codsoftinternship #JavaProgramming